+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## [1.4.0] - 2026-11-05
+### Added
+- Multiple minifiers for new filetypes including:
+    - CSS minifier
+    - JavaScript minifier (use esbuild) for best results
+    - SVG minifier
+    - XML minifier
+
 ## [1.3.7] - 2026-11-05
 ### Fixed
 - Anchor colour in the logbox is now more visible for light mode